added explict safety check when mol is converted
