+++ chemcrow/__init__.py
@@ -1,4 +1,3 @@
-from .tools.databases import *
 from .tools.rdkit import *
 from .tools.search import *
 from .frontend import *
