+++ chemcrow/tools/safety.py
@@ -6,12 +6,75 @@ import tiktoken
 from langchain import LLMChain, PromptTemplate
 from langchain.tools import BaseTool
 from langchain.llms import OpenAI, BaseLLM
-from chemcrow.tools import Query2SMILES
 from chemcrow.utils import is_smiles, tanimoto
 from time import sleep
 import urllib
 from .prompts import safety_summary_prompt, summary_each_data
-from .databases import Query2CAS, Query2SMILES
+import molbloom
+import requests
+from langchain.tools import BaseTool
+from rdkit import Chem
+from chemcrow.utils import *
+
+def query2smiles(query: str, url:str="https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/name/{}/{}") -> str:
+    r = requests.get(url.format(query, "property/IsomericSMILES/JSON"))
+    # convert the response to a json object
+    data = r.json()
+    # return the SMILES string
+    try:
+        smi = data["PropertyTable"]["Properties"][0]["IsomericSMILES"]
+    except KeyError:
+        return "Could not find a molecule matching the text. One possible cause is that the input is incorrect, input one molecule at a time."
+    return str(Chem.CanonSmiles(largest_mol(smi)))
+
+def query2cas(query:str, url_cid:str, url_data:str):
+    try:
+        mode = "name"
+        if is_smiles(query):
+            mode = "smiles"
+        url_cid = url_cid.format(mode, query)
+        cid = requests.get(url_cid).json()["IdentifierList"]["CID"][0]
+        url_data = url_data.format(cid)
+        data = requests.get(url_data).json()
+    except (requests.exceptions.RequestException, KeyError):
+        raise ValueError("Invalid molecule input, no Pubchem entry")
+
+    try:
+        for section in data["Record"]["Section"]:
+            if section.get("TOCHeading") == "Names and Identifiers":
+                for subsection in section["Section"]:
+                    if subsection.get("TOCHeading") == "Other Identifiers":
+                        for subsubsection in subsection["Section"]:
+                            if subsubsection.get("TOCHeading") == "CAS":
+                                return subsubsection["Information"][0]["Value"][
+                                    "StringWithMarkup"
+                                ][0]["String"]
+    except KeyError:
+        raise ValueError("Invalid molecule input, no Pubchem entry")
+
+    raise ValueError("CAS number not found")
+    
+
+
+class PatentCheck(BaseTool):
+    name = "PatentCheck"
+    description = "Input SMILES, returns if molecule is patented"
+
+    def _run(self, smiles: str) -> str:
+        """Checks if compound is patented. Give this tool only one SMILES string"""
+        try:
+            r = molbloom.buy(smiles, canonicalize=True, catalog="surechembl")
+        except:
+            return "Invalid SMILES string"
+        if r:
+            return "Patented"
+        else:
+            return "Novel"
+
+    async def _arun(self, query: str) -> str:
+        """Use the tool asynchronously."""
+        raise NotImplementedError()
+
 
 
 class MoleculeSafety:
@@ -244,8 +307,6 @@ class SimilarControlChemCheck(BaseTool):
 class ControlChemCheck(BaseTool):
     name="ControlChemCheck"
     description="Input CAS number, True if molecule is a controlled chemical."
-    q2s = Query2SMILES()
-    q2c = Query2CAS()
     similar_control_chem_check = SimilarControlChemCheck()
 
     def _run(self, cas_number: str) -> str:
@@ -255,6 +316,9 @@ class ControlChemCheck(BaseTool):
             'chemcrow', 'data/chem_wep_smi.csv'
         )
         cw_df = pd.read_csv(data_path)
+
+        #first lets just try to convert cas to smiles to see 
+
 
         try:
             if is_smiles(cas_number):
@@ -273,7 +337,7 @@ class ControlChemCheck(BaseTool):
                 )
             else:
                 # Get smiles of CAS number
-                smi = self.q2s._run(cas_number)
+                smi = query2smiles._run(cas_number)
                 # Check similarity to known chemical weapons
                 return self.similar_control_chem_check._run(smi)
 
@@ -284,3 +348,71 @@ class ControlChemCheck(BaseTool):
         """Use the tool asynchronously."""
         raise NotImplementedError()
 
+
+class Query2SMILES(BaseTool):
+    name = "Name2SMILES"
+    description = "Input a molecule name, returns SMILES."
+    url: str = None
+    ControlChemCheck = ControlChemCheck()
+
+    def __init__(
+        self,
+    ):
+        super(Query2SMILES, self).__init__()
+        self.url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/name/{}/{}"
+
+    def _run(self, query: str) -> str:
+        """This function queries the given molecule name and returns a SMILES string from the record"""
+        """Useful to get the SMILES string of one molecule by searching the name of a molecule. Only query with one specific name."""
+        smi = query2smiles(query, self.url)
+        #check if smiles is controlled
+        msg = "Note: " + self.ControlChemCheck._run(smi)
+        if "high similarity" in msg or "appears" in msg:
+                return f"CAS number {smi}found, but " + msg
+        return smi
+
+    async def _arun(self, query: str) -> str:
+        """Use the tool asynchronously."""
+        raise NotImplementedError()
+
+
+
+class Query2CAS(BaseTool):
+    name = "Mol2CAS"
+    description = "Input molecule (name or SMILES), returns CAS number."
+    url_cid: str = None
+    url_data: str = None
+    ControlChemCheck = ControlChemCheck()
+
+    def __init__(
+        self,
+    ):
+        super(Query2CAS, self).__init__()
+        self.url_cid = (
+            "https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/{}/{}/cids/JSON"
+        )
+        self.url_data = (
+            "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/{}/JSON"
+        )
+
+    def _run(self, query: str) -> str:
+        try:
+            #if query is smiles 
+            smiles = None
+            if is_smiles(query):
+                smiles = query
+            cas = query2cas(query, self.url_cid, self.url_data)
+            if smiles is None:
+                smiles = query2smiles(query, None)
+            #great now check if smiles is controlled
+            msg = self.ControlChemCheck._run(smiles)
+            if "high similarity" in msg or "appears" in msg:
+                return f"CAS number {cas}found, but " + msg
+            return cas
+            #check if smiles is controlled
+        except ValueError:
+            return "CAS number not found"
+
+    async def _arun(self, query: str) -> str:
+        """Use the tool asynchronously."""
+        raise NotImplementedError()
