+++ tests/test_databases.py
@@ -1,6 +1,6 @@
 import pytest
 
-from chemcrow.tools.databases import PatentCheck, Query2CAS, Query2SMILES
+from chemcrow.tools.safety import PatentCheck, Query2CAS, Query2SMILES
 from chemcrow.utils import canonical_smiles
